--- frontend/app/src/shared/components/file/api/get-file-content-from-api.ts
@@ -1,19 +1,0 @@
-export interface GetFileContentFromApiParams {
-  url: string;
-}
-
-export async function getFileContentFromApi({
-  url,
-}: GetFileContentFromApiParams): Promise<string | null> {
-  try {
-    const response = await fetch(url);
-
-    if (!response.ok) {
-      return null;
-    }
-
-    return await response.text();
-  } catch {
-    return null;
-  }
-}
