--- frontend/app/src/shared/components/file/domain/file-content.query-keys.ts
@@ -1,4 +1,0 @@
-export const fileContentQueryKeys = {
-  all: ["file-content"] as const,
-  byUrl: (url: string) => [...fileContentQueryKeys.all, url] as const,
-} as const;
