--- frontend/app/src/shared/components/file/ui/types.ts
@@ -1,5 +1,0 @@
-export interface FileViewerBaseProps {
-  url: string;
-  downloadUrl?: string;
-  fileName: string;
-}
